requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
requests==2.31.0
redis==5.0.1
celery==5.3.6
google-auth-oauthlib==1.2.0
google-auth==2.27.0
google-api-python-client==2.118.0
google-cloud-documentai==2.23.0
msal==1.26.0
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
alembic==1.12.1
openpyxl==3.1.2